+++ gsw_xarray/_attributes.py
